Remove analysis_utils.py
