+++ src/kbmod/__init__.py
@@ -5,9 +5,12 @@ try:
 except ImportError:
     warnings.warn("Unable to determine the package version. " "This is likely a broken installation.")
 
+# This is needed for compatibility with some older compilers: erfinv needs to be
+# imported before other packages though I'm not sure why.
+from scipy.special import erfinv
+
 from . import (
     analysis,
-    analysis_utils,
     data_interface,
     file_utils,
     filters,
