+++ src/kbmod/filters/__init__.py
@@ -1,1 +1,7 @@
-from . import base_filter, stamp_filters, stats_filters
+from . import (
+    base_filter,
+    clustering_filters,
+    sigma_g_filter,
+    stamp_filters,
+    stats_filters,
+)
