+++ tests/test_filtering.py
@@ -3,7 +3,6 @@ import unittest
 
 import numpy as np
 
-from kbmod.analysis_utils import PostProcess
 from kbmod.search import *
