+++ tests/test_search.py
@@ -2,7 +2,9 @@ import unittest
 
 import numpy as np
 
-from kbmod.fake_data.fake_data_creator import add_fake_object, make_fake_layered_image
+from kbmod.configuration import SearchConfiguration
+from kbmod.fake_data.fake_data_creator import add_fake_object, make_fake_layered_image, FakeDataSet
+from kbmod.run_search import SearchRunner
 from kbmod.search import *
 from kbmod.trajectory_generator import KBMODV1Search
 from kbmod.trajectory_utils import make_trajectory
@@ -132,6 +134,54 @@ class test_search(unittest.TestCase):
         self.search.clear_results()
         results = self.search.get_results(0, 10)
         self.assertEqual(len(results), 0)
+
+    def test_load_and_filter_results_lh(self):
+        time_list = [i / self.img_count for i in range(self.img_count)]
+        fake_ds = FakeDataSet(
+            self.dim_x,
+            self.dim_y,
+            time_list,
+            noise_level=1.0,
+            psf_val=0.5,
+            use_seed=True,
+        )
+
+        # Create fake result trajectories with given initial likelihoods. The 1st is
+        # filtered by max likelihood. The 4th and 5th are filtered by min likelihood.
+        trjs = [
+            make_trajectory(20, 20, 0, 0, 500.0, 9000.0, self.img_count),
+            make_trajectory(30, 30, 0, 0, 100.0, 100.0, self.img_count),
+            make_trajectory(40, 40, 0, 0, 50.0, 50.0, self.img_count),
+            make_trajectory(50, 50, 0, 0, 1.0, 2.0, self.img_count),
+            make_trajectory(60, 60, 0, 0, 1.0, 1.0, self.img_count),
+        ]
+        for trj in trjs:
+            fake_ds.insert_object(trj)
+
+        # Create the stack search and insert the fake results.
+        search = StackSearch(fake_ds.stack)
+        search.set_results(trjs)
+
+        # Do the loading and filtering.
+        config = SearchConfiguration()
+        overrides = {
+            "clip_negative": False,
+            "chunk_size": 500000,
+            "lh_level": 10.0,
+            "max_lh": 1000.0,
+            "num_cores": 1,
+            "num_obs": 5,
+            "sigmaG_lims": [25, 75],
+        }
+        config.set_multiple(overrides)
+
+        runner = SearchRunner()
+        results = runner.load_and_filter_results(search, config)
+
+        # Only two of the middle results should pass the filtering.
+        self.assertEqual(results.num_results(), 2)
+        self.assertEqual(results.results[0].trajectory.y, 30)
+        self.assertEqual(results.results[1].trajectory.y, 40)
 
     @unittest.skipIf(not HAS_GPU, "Skipping test (no GPU detected)")
     def test_evaluate_single_trajectory(self):
